+++ src/Services/Catalogs/Ecommerce.Services.Catalog.Application/Features/Products/Commands/UpdateProduct/UpdateProductCommand.cs
@@ -13,8 +13,8 @@ public record UpdateProductCommand(
     List<string> ImageUrls,
     long? CategoryId,
     string? AttributesJson,
-    double Weight = 0,
-    double Length = 0,
-    double Width = 0,
-    double Height = 0
+    int Weight = 0,
+    int Length = 0,
+    int Width = 0,
+    int Height = 0
 ) : ICommand<ProductResponse>;
