+++ src/Services/Catalogs/Ecommerce.Services.Catalog.Application/Features/Products/Dtos/VariantDto.cs
@@ -17,9 +17,9 @@ public class VariantDto
     public decimal DiscountPrice { get; set; }
     public string VariantName { get; set; } = string.Empty;
     public long ShopId { get; set; }
-    public double Weight { get; set; }
-    public double Length { get; set; }
-    public double Width { get; set; }
-    public double Height { get; set; }
+    public int Weight { get; set; }
+    public int Length { get; set; }
+    public int Width { get; set; }
+    public int Height { get; set; }
     public string ThumbnailUrl { get; set; } = string.Empty;
 }
